Move UI handlebars service in the good folder
