+++ ui/main/src/app/business/services/card/handlebars.service.ts
