+++ ui/main/src/app/modules/usercard/usercard.component.ts
@@ -17,7 +17,7 @@ import {Guid} from 'guid-typescript';
 import {DomSanitizer, SafeHtml} from '@angular/platform-browser';
 import {EntitiesService} from 'app/business/services/entities.service';
 import {ProcessesService} from 'app/business/services/processes.service';
-import {HandlebarsService} from '../card/services/handlebars.service';
+import {HandlebarsService} from '../../business/services/card/handlebars.service';
 import {DetailContext} from '@ofModel/detail-context.model';
 import {map} from 'rxjs/operators';
 import {MessageLevel} from '@ofModel/message.model';
